Speed up npm test: parallelize workspaces and split slow vitest file
- Root `test` now builds @mcpjam/sdk once, runs sdk/cli/design-system/
  inspector/mcp tests in parallel via concurrently, then runs sdk
  packaging at the end (it rebuilds dist and would race with parallel
  consumers).
- Add `test:fast` and `typecheck:fast` in @mcpjam/mcp that skip the
  redundant sdk rebuild (the root pipeline already built it). Standalone
  `npm test -w @mcpjam/mcp` still rebuilds for safety.
- Split sdk/tests/MCPClientManager.test.ts: extract the `server management`
  (~50s, 13 tests) and `multiple servers` (~12s, 3 tests) describe blocks
  into their own files so vitest's worker pool runs them in parallel
  instead of serially within one file. No test logic changes.

Wall time on this machine: ~127s -> ~72s (43% faster).
