+++ package.json
@@ -15,7 +15,8 @@
     "build:inspector": "npm run build -w @mcpjam/inspector",
     "typecheck": "npm run typecheck -w @mcpjam/sdk && npm run build -w @mcpjam/sdk && npm run typecheck -w @mcpjam/cli && npm run typecheck -w @mcpjam/design-system && npm run typecheck -w @mcpjam/mcp",
     "check:mcp-v1-runtime-imports": "! rg -n '@modelcontextprotocol/sdk' sdk/src cli/src mcpjam-inspector/client/src mcpjam-inspector/server -g '!**/dist/**'",
-    "test": "npm run check:mcp-v1-runtime-imports && npm run test:ordered",
+    "test": "npm run check:mcp-v1-runtime-imports && npm run build -w @mcpjam/sdk && npm run test:parallel && npm run test:packaging -w @mcpjam/sdk",
+    "test:parallel": "concurrently -n sdk,cli,design,inspector,mcp,mcp-tc -c blue,green,yellow,magenta,cyan,red --kill-others-on-fail \"npm run test -w @mcpjam/sdk\" \"npm run test -w @mcpjam/cli\" \"npm run test -w @mcpjam/design-system\" \"npm run test -w @mcpjam/inspector\" \"npm run test:fast -w @mcpjam/mcp\" \"npm run typecheck:fast -w @mcpjam/mcp\"",
     "test:ordered": "npm run test -w @mcpjam/sdk && npm run test:packaging -w @mcpjam/sdk && npm run test -w @mcpjam/cli && npm run test -w @mcpjam/design-system && npm run test -w @mcpjam/inspector && npm run test -w @mcpjam/mcp && npm run typecheck -w @mcpjam/mcp",
     "verify": "npm run typecheck && npm run test && npm run build:inspector",
     "release:status": "changeset status",
